+++ montecarlo/montecarlo.h
@@ -2,11 +2,12 @@
 #include "../histogram/histogram.h"
 typedef struct montecarlo{
 	histogram* histo;
+	//histogram* histo...;
 	state* state;
 } montecarlo;
 
 
-montecarlo* montecarloInit(double initial_dispersion, unsigned int histogram_size, unsigned int histogram_range);
+montecarlo* montecarloInit(double initial_dispersion, unsigned int histogram_size, double histogram_range);
 
 void runOneStep(montecarlo* mc);
